requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --frozen --output-file=requirements.txt --no-dev --no-hashes
-e .
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama==0.4.6 ; platform_system == 'Windows'
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.51.0
deprecated==1.2.18
docker==7.1.0
fastapi==0.115.12
filelock==3.18.0
flask==3.1.0
fonttools==4.57.0
fsspec==2025.3.0
gitdb==4.0.12
gitpython==3.1.44
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.1 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
gunicorn==23.0.0 ; platform_system != 'Windows'
h11==0.16.0
idna==3.10
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
mako==1.3.10
markdown==3.8
markupsafe==3.0.2
matplotlib==3.10.1
mlflow==2.22.0
mlflow-skinny==2.22.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cuda-cupti-cu12==12.1.105 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cuda-nvrtc-cu12==12.1.105 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cuda-runtime-cu12==12.1.105 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cudnn-cu12==8.9.2.26 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cufft-cu12==11.0.2.54 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-curand-cu12==10.3.2.106 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cusolver-cu12==11.4.5.107 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-cusparse-cu12==12.1.0.106 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-nccl-cu12==2.19.3 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-nvjitlink-cu12==12.4.127 ; platform_machine == 'x86_64' and platform_system == 'Linux'
nvidia-nvtx-cu12==12.1.105 ; platform_machine == 'x86_64' and platform_system == 'Linux'
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging==24.2
pandas==2.2.3
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.30.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pyparsing==3.2.3
pysoundfile==0.9.0.post1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310 ; sys_platform == 'win32'
pyyaml==6.0.2
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
sqlalchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.67.1
triton==2.2.0 ; python_full_version < '3.12' and platform_machine == 'x86_64' and platform_system == 'Linux'
typing-extensions==4.12.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
waitress==3.0.2 ; platform_system == 'Windows'
werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
